refactor: chipGroup 리팩토링
