+++ packages/mobile/src/components/molecules/ChipGroup/index.tsx
@@ -1,9 +1,39 @@
-import Chip from "src/components/atoms/Chip";
+import { NavLink } from "react-router-dom";
 
-function ChipGroup() {
-  // TODO: 여러개의 칩을 나타냅니다.
+import $ from "./style.module.scss";
 
-  return <Chip />;
+type ItemProps = {
+  id: number;
+  name: string;
+  path: string;
+};
+
+type Props = {
+  list: ItemProps[];
+  handleSelectMenu: (idx: number) => void;
+  menuType: number;
+};
+
+function ChipGroup({ list, handleSelectMenu, menuType }: Props) {
+  return (
+    <div className={$.menu}>
+      <div className={$.list}>
+        {list.map((item, idx) => {
+          return (
+            <NavLink
+              key={`menu-${item.id}`}
+              to={item.path}
+              className={$["menu-link"]}
+              onClick={() => handleSelectMenu(idx)}
+              aria-selected={menuType === idx + 1}
+            >
+              {item.name}
+            </NavLink>
+          );
+        })}
+      </div>
+    </div>
+  );
 }
 
 export default ChipGroup;
