+++ packages/mobile/src/components/molecules/ChipGroup/style.module.scss
@@ -1,0 +1,52 @@
+@import "@shared/styles/color.scss";
+@import "@shared/styles/variables.scss";
+@import "@shared/styles/mixin.scss";
+
+.menu {
+  position: fixed;
+  top: 60px;
+  left: 0;
+  right: 0;
+  z-index: 100;
+  padding: 12px 21px 12px 13px;
+  background-color: $white;
+  border-top: 3px solid $background;
+}
+
+.list {
+  display: flex;
+
+  @include flicking;
+}
+
+.menu-link {
+  display: block;
+  font-size: 14px;
+  line-height: 18px;
+  padding: 8px 12px;
+  white-space: nowrap;
+  color: $black-80;
+  border-radius: 100px;
+
+  &:not(:first-child) {
+    margin-left: 16px;
+  }
+
+  &[aria-selected="true"] {
+    background-color: $red-20;
+    color: $red-100;
+
+    &:hover {
+      @include hover-status($red-100, none);
+    }
+  }
+
+  &[aria-selected="false"] {
+    background-color: $black-10;
+    color: $black-80;
+
+    &:hover {
+      @include hover-status($black-80, none);
+    }
+  }
+}
