--- packages/mobile/src/components/molecules/Chips/style.module.scss
@@ -1,49 +1,0 @@
-@import "@shared/styles/color.scss";
-@import "@shared/styles/variables.scss";
-@import "@shared/styles/mixin.scss";
-
-.hash {
-  position: fixed;
-  top: 121px;
-  left: 0;
-  right: 0;
-  display: flex;
-  padding: 11px 15px 11px 13px;
-  background-color: $white;
-  border-top: 3px solid $background;
-  box-shadow: $border-box-shadow;
-
-  @include flicking;
-}
-
-.hash-link {
-  display: block;
-  font-size: 14px;
-  line-height: 18px;
-  padding: 7px 8px 7px 10px;
-  white-space: nowrap;
-
-  &:not(:first-child) {
-    margin-left: 12px;
-  }
-
-  &.active {
-    background-color: $red-20;
-    color: $red-100;
-    border-radius: 100px;
-
-    &:hover {
-      @include hover-status($red-100, none);
-    }
-  }
-
-  &[aria-selected="false"] {
-    background-color: $black-10;
-    color: $black-80;
-    border-radius: 100px;
-
-    &:hover {
-      @include hover-status($black-80, none);
-    }
-  }
-}
